// src/domain/models/organization-model.ts
export type OrganizationModel = {
  cidade?: string;
  cidade_id?: string;
  cidade_nome?: string;
  id?: string;
  name?: string;
  tipo_entidade_id: string;
  tipo_entidade_nome: string;
  cnpj: string;
  razao_social: string;
  nome_fantasia: string;
  logradouro: string;
  numero: string;
  complemento?: string;
  bairro: string;
  cep: string;
  telefone?: string;
  email: string;
  site: string;
  ativo?: string;
  created_at?: string;
  created_by?: string;
  updated_at?: string;
  updated_by?: string;
  logo?: string;
  representante?: string;
  representante_cargo?: string;
  tce_seguradora_id?: string;
  tce_num_apolice?: string;
  grupo?: string;
  tce_modelo_rescisao?: string;
  codigo_integracao?: string;
  tce_Seguradora?: string;
  tce_seguradora_nome?: string;
};